+++ cogs/daily_reminder.py
@@ -1,39 +1,10 @@
 import discord
 from discord.ext import commands, tasks
 
-class reminders(commands.Cog):
+class daily_reminder(commands.Cog):
 
   def __init__(self, client):
     self.client = client
-
-  @commands.command(aliases=["set-alarm"])
-  async def alarm(self, ctx, time):
-      """Sets an alarm for a given time
-      Args:
-          ctx (discord.ext.commands.Context): Represents the context in which a command is being invoked under.
-          time ([type]): time in 24 hour format
-      """
-      # Take user input (time in 24-hour format), for example 23:00
-      # Check if it is a valid 24-hour format, and return an apppropriate message
-      # Parse the user input using time.split(":")
-      # Remind the user as soon as it's time
-      pass
-
-
-
-  @commands.command(aliases=["hourly-reminder","set-hourly-reminder"])
-  async def hourly_reminder(self, ctx, task):
-      """Sets a reminder for a give task each hour
-      Args:
-          ctx (discord.ext.commands.Context): Represents the context in which a command is being invoked under.
-          task (str): The task to be reminded of
-      """
-      # Takes input from the user (task) about what they would like to accomplish
-      # Send a reminder after one hour to see how far they have come up with the task
-      # At the end of every hour, ask the user if they have completed the task.
-      # Wait for the stop command from user (You may use "client.wait_for")
-      pass
-
 
   @commands.command(aliases=["daily-reminder", "set-daily-alarm"])
   async def daily_reminder(self, ctx,*, task):
@@ -98,4 +69,4 @@ class reminders(commands.Cog):
       self.reminder_loop.cancel()
 
 def setup(client):
-  client.add_cog(reminders(client))
+  client.add_cog(daily_reminder(client))
